kvconnector v0 deprecation
