+++ .github/mergify.yml
@@ -310,7 +310,7 @@ pull_request_rules:
       - files~=^examples/online_serving/disaggregated[^/]*/.*
       - files~=^examples/offline_inference/disaggregated[^/]*/.*
       - files~=^examples/others/lmcache/
-      - files~=^tests/v1/kv_connector/
+      - files~=^tests/kv_connector/
       - files~=^vllm/distributed/kv_transfer/
       - title~=(?i)\bP/?D\b
       - title~=(?i)NIXL
