+++ tests/kv_connector/nixl_integration/run_tpu_disagg_accuracy_test.sh
@@ -21,7 +21,7 @@ BLOCK_SIZE=${BLOCK_SIZE:-32}
 
 # execution env
 GIT_ROOT=$(git rev-parse --show-toplevel)
-EXP_ROOT="${GIT_ROOT}/tests/v1/kv_connector/nixl_integration"
+EXP_ROOT="${GIT_ROOT}/tests/kv_connector/nixl_integration"
 CONDA_PATH=${CONDA_PATH:-"/home/${USER}/anaconda3"}
 CONDA_ENV_NAME=${CONDA_ENV_NAME:-"nixl"}
