+++ vllm/distributed/kv_transfer/__init__.py
@@ -2,12 +2,12 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 from vllm.distributed.kv_transfer.kv_transfer_state import (
-    KVConnectorBaseType, ensure_kv_transfer_initialized,
+    KVConnectorBase, ensure_kv_transfer_initialized,
     ensure_kv_transfer_shutdown, get_kv_transfer_group, has_kv_transfer_group,
     is_v1_kv_transfer_group)
 
 __all__ = [
     "get_kv_transfer_group", "has_kv_transfer_group",
     "is_v1_kv_transfer_group", "ensure_kv_transfer_initialized",
-    "ensure_kv_transfer_shutdown", "KVConnectorBaseType"
+    "ensure_kv_transfer_shutdown", "KVConnectorBase"
 ]
