+++ vllm/distributed/kv_transfer/kv_connector/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from vllm.distributed.kv_transfer.kv_connector.base import (KVConnectorBase,
+                                                            KVConnectorRole)
+
+__all__ = ["KVConnectorRole", "KVConnectorBase"]
