+++ vllm/distributed/kv_transfer/kv_connector/factory.py
@@ -6,9 +6,8 @@ from typing import TYPE_CHECKING, Callable
 
 # yapf: disable
 import vllm.envs as envs
-from vllm.distributed.kv_transfer.kv_connector.base import (
-    KVConnectorBase, KVConnectorBaseType)
-from vllm.distributed.kv_transfer.kv_connector.v1 import KVConnectorRole
+from vllm.distributed.kv_transfer.kv_connector import KVConnectorRole
+from vllm.distributed.kv_transfer.kv_connector.base import KVConnectorBase
 from vllm.logger import init_logger
 
 # yapf: enable
@@ -62,8 +61,8 @@ class KVConnectorFactory:
 
     @classmethod
     def get_connector_class(
-            cls, kv_transfer_config: "KVTransferConfig"
-    ) -> type[KVConnectorBaseType]:
+            cls,
+            kv_transfer_config: "KVTransferConfig") -> type[KVConnectorBase]:
         """Get the connector class by name."""
         connector_name = kv_transfer_config.kv_connector
         if connector_name in cls._registry:
@@ -84,30 +83,30 @@ class KVConnectorFactory:
 
 KVConnectorFactory.register_connector(
     "SharedStorageConnector",
-    "vllm.distributed.kv_transfer.kv_connector.v1.shared_storage_connector",
+    "vllm.distributed.kv_transfer.kv_connector.shared_storage_connector",
     "SharedStorageConnector")
 
 KVConnectorFactory.register_connector(
     "P2pNcclConnector",
-    "vllm.distributed.kv_transfer.kv_connector.v1.p2p.p2p_nccl_connector",
+    "vllm.distributed.kv_transfer.kv_connector.p2p.p2p_nccl_connector",
     "P2pNcclConnector")
 
 KVConnectorFactory.register_connector(
     "LMCacheConnectorV1",
-    "vllm.distributed.kv_transfer.kv_connector.v1.lmcache_connector",
+    "vllm.distributed.kv_transfer.kv_connector.lmcache_connector",
     "LMCacheConnectorV1")
 
 KVConnectorFactory.register_connector(
     "NixlConnector",
-    "vllm.distributed.kv_transfer.kv_connector.v1.nixl_connector",
+    "vllm.distributed.kv_transfer.kv_connector.nixl_connector",
     "NixlConnector")
 
 KVConnectorFactory.register_connector(
     "MultiConnector",
-    "vllm.distributed.kv_transfer.kv_connector.v1.multi_connector",
+    "vllm.distributed.kv_transfer.kv_connector.multi_connector",
     "MultiConnector")
 
 KVConnectorFactory.register_connector(
     "OffloadingConnector",
-    "vllm.distributed.kv_transfer.kv_connector.v1.offloading_connector",
+    "vllm.distributed.kv_transfer.kv_connector.offloading_connector",
     "OffloadingConnector")
