+++ vllm/distributed/kv_transfer/kv_connector/p2p/p2p_nccl_connector.py
@@ -8,9 +8,9 @@ import regex as re
 import torch
 
 from vllm.config import VllmConfig
-from vllm.distributed.kv_transfer.kv_connector.v1.base import (
-    KVConnectorBase_V1, KVConnectorMetadata, KVConnectorRole)
-from vllm.distributed.kv_transfer.kv_connector.v1.p2p.p2p_nccl_engine import (
+from vllm.distributed.kv_transfer.kv_connector.base import (
+    KVConnectorBase, KVConnectorMetadata, KVConnectorRole)
+from vllm.distributed.kv_transfer.kv_connector.p2p.p2p_nccl_engine import (
     P2pNcclEngine)
 from vllm.distributed.parallel_state import get_world_group
 from vllm.logger import init_logger
@@ -64,7 +64,7 @@ class P2pNcclConnectorMetadata(KVConnectorMetadata):
             ReqMeta.make_meta(request_id, token_ids, block_ids, block_size))
 
 
-class P2pNcclConnector(KVConnectorBase_V1):
+class P2pNcclConnector(KVConnectorBase):
 
     def __init__(self, vllm_config: "VllmConfig", role: KVConnectorRole):
         super().__init__(vllm_config=vllm_config, role=role)
