--- vllm/distributed/kv_transfer/kv_connector/v1/__init__.py
@@ -1,6 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from vllm.distributed.kv_transfer.kv_connector.v1.base import (
-    KVConnectorBase_V1, KVConnectorRole)
-
-__all__ = ["KVConnectorRole", "KVConnectorBase_V1"]
