+++ vllm/distributed/kv_transfer/kv_transfer_state.py
@@ -3,19 +3,18 @@
 from typing import TYPE_CHECKING, Optional
 
 from vllm import envs
-from vllm.distributed.kv_transfer.kv_connector.base import KVConnectorBaseType
+from vllm.distributed.kv_transfer.kv_connector import (KVConnectorBase,
+                                                       KVConnectorRole)
 from vllm.distributed.kv_transfer.kv_connector.factory import (
     KVConnectorFactory)
-from vllm.distributed.kv_transfer.kv_connector.v1 import (KVConnectorBase_V1,
-                                                          KVConnectorRole)
 
 if TYPE_CHECKING:
     from vllm.config import VllmConfig
 
-_KV_CONNECTOR_AGENT: Optional[KVConnectorBaseType] = None
+_KV_CONNECTOR_AGENT: Optional[KVConnectorBase] = None
 
 
-def get_kv_transfer_group() -> KVConnectorBaseType:
+def get_kv_transfer_group() -> KVConnectorBase:
     assert _KV_CONNECTOR_AGENT is not None, (
         "disaggregated KV cache transfer parallel group is not initialized")
     return _KV_CONNECTOR_AGENT
@@ -26,7 +25,7 @@ def has_kv_transfer_group() -> bool:
 
 
 def is_v1_kv_transfer_group(
-        connector: Optional[KVConnectorBaseType] = None) -> bool:
+        connector: Optional[KVConnectorBase] = None) -> bool:
     """Check if the KV connector is the v1 connector.
     If the argument is None, it will check the global KV connector
 
@@ -44,7 +43,7 @@ def is_v1_kv_transfer_group(
     if connector is None:
         return False
 
-    return isinstance(connector, KVConnectorBase_V1)
+    return isinstance(connector, KVConnectorBase)
 
 
 def ensure_kv_transfer_initialized(vllm_config: "VllmConfig") -> None:
