+++ vllm/v1/metrics/loggers.py
@@ -9,7 +9,7 @@ from typing import Callable, Optional, Union
 import prometheus_client
 
 from vllm.config import SupportsMetricsInfo, VllmConfig
-from vllm.distributed.kv_transfer.kv_connector.v1.metrics import (
+from vllm.distributed.kv_transfer.kv_connector.metrics import (
     KVConnectorLogging)
 from vllm.logger import init_logger
 from vllm.v1.core.kv_cache_utils import PrefixCachingMetrics
